+++ components/Search/Search.module.scss
@@ -1,94 +1,99 @@
 @import "@/styles/const.scss";
 .notice__wrapper {
-    width: 100%;
-    background-color: var(--white);
+  width: 100%;
+  background-color: var(--white);
 }
 
 .notice__container {
-    position: relative;
-    max-width: 964px;
-    height: 100%;
-    margin: 0 auto;
-    padding: 60px 0;
-    display: flex;
-    flex-direction: column;
-    gap: 35px;
+  position: relative;
+  max-width: 964px;
+  height: 100%;
+  margin: 0 auto;
+  padding: 60px 0;
+  display: flex;
+  flex-direction: column;
+  gap: 35px;
 }
 
 .noticeTitle__container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
 }
 
 .noticeTitle__options {
-    display: flex;
-    align-items: center;
-    gap: 10px;
+  display: flex;
+  align-items: center;
+  gap: 10px;
 }
 
 .title {
-    font-size: 28px;
-    font-weight: 600;
+  font-size: 28px;
+  font-weight: 600;
 }
 
 .title_keyword {
-    color: var(--primary);
+  color: var(--primary);
 }
 
 .post__grid {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-gap: 14px;
-    justify-items: center;
-    width: 100%;
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-gap: 14px;
+  justify-items: center;
+  width: 100%;
 }
 
 .filter__btn {
-    width: 79px;
-    height: 30px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: var(--red30);
-    color: var(--white);
-    font-size: 14px;
-    font-weight: 600;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
+  width: 79px;
+  height: 30px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: var(--red30);
+  color: var(--white);
+  font-size: 14px;
+  font-weight: 600;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
 }
 
 .filter__wrapper {
-    position: relative;
+  position: relative;
+}
+
+.empty-result {
+  align-self: center;
+  margin-top: 200px;
 }
 
 @media (max-width: $TABLET) {
-    .notice__container {
-        margin: 0 32px;
-    }
-    .post__grid {
-        grid-template-columns: repeat(2, 1fr);
-        grid-gap: 28px 0;
-    }
+  .notice__container {
+    margin: 0 32px;
+  }
+  .post__grid {
+    grid-template-columns: repeat(2, 1fr);
+    grid-gap: 28px 0;
+  }
 }
 
 @media (max-width: $MOBILE) {
-    .notice__container {
-        padding: 40px 12px;
-        margin: 0;
-        width: 100%;
-    }
-    .post__grid {
-        grid-template-columns: repeat(2, 1fr);
-        width: 100%;
-    }
-    .noticeTitle__container {
-        flex-direction: column;
-        align-items: flex-start;
-        grid-gap: 10px;
-    }
-    .title {
-        font-size: 20px;
-    }
+  .notice__container {
+    padding: 40px 12px;
+    margin: 0;
+    width: 100%;
+  }
+  .post__grid {
+    grid-template-columns: repeat(2, 1fr);
+    width: 100%;
+  }
+  .noticeTitle__container {
+    flex-direction: column;
+    align-items: flex-start;
+    grid-gap: 10px;
+  }
+  .title {
+    font-size: 20px;
+  }
 }
